Move component-level derived data to a memoized function in the component
